feat(inkspill): check win

+++ common/src/events.rs
@@ -2,5 +2,5 @@ use bevy::prelude::Event;
 
 #[derive(Event)]
 pub struct EndGame {
-    pub score: u32,
+    pub score: usize,
 }
